UR_Atmosphere: changed sqrt to sqrtf on CaSLSI atm definitions.

+++ libraries/UR_Atmosphere/atmosphere_definitions.h
@@ -80,7 +80,7 @@ namespace ISA_MATH_CONST {
     const float CGamma     = 1.4f;    // [-]
     const float CGammaRGas = (CGamma * CRGasSI) / (CftTOm * CftTOm);    // [ft^2/(s^2*K)]
 
-    const float CaSLSI         = sqrt(CGamma * CRGasSI * T0);
+    const float CaSLSI         = sqrtf(CGamma * CRGasSI * T0);
     const float CPressureSLSI  = 101325.0f;                             // [Pa] = [N/m^2]
     const float CaSLNU         = CaSLSI * CmPsTOkn;                     // [kts] Nautical Unit
     const float CRhoSLSI       = 1.225f;                                // [kg/m^3]
